requirements.txt
aiogram==3.0.0b6
asyncio
python-decouple
apscheduler